When building packages pass option '--ownership preserve-other'.
In some cases, it is important to preserve a different default
group and permissions for a file or directory. For example,
in mutt package, the file mutt_dotlock should be g+s and owned
by root:mail. Prior to this patch, it was impossible to do this.

So now we're going to use '--ownership preserve-other' when
creating packages, it works like this: if a file/directory
has different attributes than the current user, it will be
preserve in the package, all other files (with ownership
by the build user)  will be changed to what pkgbuild thinks is the
default for the system (say root:wheel).
The old behavior was --ownership preserve.

For more information, see pkgbuild(1).

Signed-off-by: Rudá Moura <ruda.moura@gmail.com>