mtr: Binary in /usr/local/bin/mtr and setuid.
Following the KISS principle here.

As /usr/local/sbin is not available by default in PATH,
I'm going to change mtr to go to /usr/local/bin.

Also, because OS X has traceroute set as "+s" (setuid),
we're going to set mtr as setuid too. It's less secure,
but it's easy for users to use, they don't need to run with sudo.

Thanks to Melvin Mughal	for raising these issues.

Signed-off-by: Rudá Moura <ruda.moura@gmail.com>